Added initial dashkey firmware code dump
Note: This code is *very incomplete* in terms of compatibility and
testing with the Teensy used in openclack; it requires a number of
changes and improvements to support that controller (it currently
targets an atmega32u2 based design with a custom layout) so use at your
own peril (for the moment).

Eventually this code will include easy driver support for multiple
controllers and key layouts and should be compatible with a number of
avr-based usb enabled microcontrollers. The code depends on the most
excellent LUFA library; notes are included for obtaining a copy and
setting up a build environment (subject to change).