Added support for openclack pinout/layout
The dashkey header file now includes an OPENCLACK define that selects
between the openclack layout and the dashkey-custom layout this firmware
is originally derived from. The code as-is has been tested to flash and
run correctly on an openclack board (and it enumerates the USB devices
correctly) but doesn't seem to be passing keypresses or LEDs. Needs
further investigation.